+++ src/support/db/users.py
@@ -7,7 +7,7 @@ users = sa.Table(
     metadata,
 
     sa.Column("db_id", sa.Integer, primary_key=True, unique=True),
-    sa.Column("id", sa.Integer, index=True, unique=True, nullable=False),
+    sa.Column("id", sa.BigInteger, index=True, unique=True, nullable=False),
     sa.Column("name", sa.String, nullable=False),
     sa.Column("username", sa.String, default=None),
